+++ pygqlc/__init__.py
@@ -1,3 +1,4 @@
+from .__version__ import __version__
 from .GraphQLClient import GraphQLClient, GQLResponseException
 from .QueryParser import QueryParser
 from .MutationParser import MutationParser
@@ -6,5 +7,3 @@ from .helper_modules.Singleton import Singleton
 
 # * Package name:
 name = 'pygqlc'
-
-from .__version__ import __version__